refactor: Home page transformed into job listings page.
