+++ src/pages/Jobs/style.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const PageBody = styled.div`
+  padding: 50px;
+  background-color: #0c2240;
+  display: flex;
+  height: 100vh;
+  padding: 32px 160px 0 160px;
+  gap: 50px;
+`
